+++ packages/fluentui/react-bindings/src/hooks/useMergedRefs.ts
@@ -13,23 +13,23 @@ export type RefObjectFunction<T> = React.RefObject<T> & ((value: T) => void);
  * @returns A function with an attached "current" prop, so that it can be treated like a RefObject.
  */
 export function useMergedRefs<T>(...refs: (React.Ref<T> | undefined)[]): RefObjectFunction<T> {
-  const mergedCallback: RefObjectFunction<T> = (React.useCallback(
+  const mergedCallback: RefObjectFunction<T> = React.useCallback(
     (value: T) => {
       // Update the "current" prop hanging on the function.
-      ((mergedCallback as unknown) as React.MutableRefObject<T>).current = value;
+      (mergedCallback as unknown as React.MutableRefObject<T>).current = value;
 
       for (const ref of refs) {
         if (typeof ref === 'function') {
           ref(value);
         } else if (ref) {
           // work around the immutability of the React.Ref type
-          ((ref as unknown) as React.MutableRefObject<T>).current = value;
+          (ref as unknown as React.MutableRefObject<T>).current = value;
         }
       }
     },
     // eslint-disable-next-line react-hooks/exhaustive-deps -- already exhaustive
     [...refs],
-  ) as unknown) as RefObjectFunction<T>;
+  ) as unknown as RefObjectFunction<T>;
 
   return mergedCallback;
 }