+++ packages/react-components/react-progress/src/components/ProgressBar/useProgressBarStyles.ts
@@ -42,11 +42,11 @@ const useRootStyles = makeStyles({
   root: {
     display: 'block',
     backgroundColor: tokens.colorNeutralBackground6,
-    justifySelf: 'stretch',
+    width: '100%',
     ...shorthands.overflow('hidden'),
 
     '@media screen and (forced-colors: active)': {
-      ...shorthands.borderBottom('1px', 'solid', 'CanvasText'),
+      backgroundColor: 'CanvasText',
     },
   },
   rounded: {
@@ -68,13 +68,11 @@ const useRootStyles = makeStyles({
  */
 const useBarStyles = makeStyles({
   base: {
+    '@media screen and (forced-colors: active)': {
+      backgroundColor: 'Highlight',
+    },
     ...shorthands.borderRadius('inherit'),
-  },
-  medium: {
-    height: barThicknessValues.medium,
-  },
-  large: {
-    height: barThicknessValues.large,
+    height: '100%',
   },
   nonZeroDeterminate: {
     transitionProperty: 'width',
@@ -105,9 +103,6 @@ const useBarStyles = makeStyles({
 
   brand: {
     backgroundColor: tokens.colorCompoundBrandBackground,
-    '@media screen and (forced-colors: active)': {
-      backgroundColor: 'Highlight',
-    },
   },
 
   error: {
@@ -145,7 +140,6 @@ export const useProgressBarStyles_unstable = (state: ProgressBarState): Progress
       barStyles.brand,
       value === undefined && barStyles.indeterminate,
       value === undefined && dir === 'rtl' && barStyles.rtl,
-      barStyles[thickness],
       value !== undefined && value > ZERO_THRESHOLD && barStyles.nonZeroDeterminate,
       color && value !== undefined && barStyles[color],
       state.bar.className,