+++ packages/react-components/react-provider/src/components/FluentProvider/useFluentProviderContextValues.ts
@@ -3,13 +3,26 @@ import type { FluentProviderContextValues, FluentProviderState } from './FluentP
 
 export function useFluentProviderContextValues_unstable(state: FluentProviderState): FluentProviderContextValues {
   // eslint-disable-next-line @typescript-eslint/naming-convention
-  const { applyStylesToPortals, dir, root, targetDocument, theme, themeClassName, overrides_unstable } = state;
+  const {
+    applyStylesToPortals,
+    // eslint-disable-next-line @typescript-eslint/naming-convention
+    customStyleHooks_unstable,
+    dir,
+    root,
+    targetDocument,
+    theme,
+    themeClassName,
+    // eslint-disable-next-line @typescript-eslint/naming-convention
+    overrides_unstable,
+  } = state;
 
   const provider = React.useMemo(() => ({ dir, targetDocument }), [dir, targetDocument]);
   // "Tooltip" component mutates an object in this context, instance should be stable
   const [tooltip] = React.useState(() => ({}));
 
   return {
+    // eslint-disable-next-line @typescript-eslint/naming-convention
+    customStyleHooks_unstable,
     // eslint-disable-next-line @typescript-eslint/naming-convention
     overrides_unstable,
     provider,