+++ packages/react-components/react-skeleton/src/components/Skeleton/index.ts
@@ -2,4 +2,5 @@ export * from './Skeleton';
 export * from './Skeleton.types';
 export * from './renderSkeleton';
 export * from './useSkeleton';
+export * from './useSkeletonContextValues';
 export * from './useSkeletonStyles';