+++ packages/react-examples/src/react/Pickers/Pickers.doc.tsx
@@ -6,10 +6,14 @@ import { IDocPageProps } from '@fluentui/react/lib/common/DocPage.types';
 import { TagPickerBasicExample } from './TagPicker.Basic.Example';
 import { TagPickerInlineExample } from './TagPicker.Inline.Example';
 
-const TagPickerExampleCode = require('!raw-loader?esModule=false!@fluentui/react-examples/src/react/Pickers/TagPicker.Basic.Example.tsx') as string;
-const TagPickerInlineExampleCode = require('!raw-loader?esModule=false!@fluentui/react-examples/src/react/Pickers/TagPicker.Inline.Example.tsx') as string;
-const PickerCustomResultExampleCode = require('!raw-loader?esModule=false!@fluentui/react-examples/src/react/Pickers/Picker.CustomResult.Example.tsx') as string;
-const TagPickerCustomRemoveIconExampleCode = require('!raw-loader?esModule=false!@fluentui/react-examples/src/react/Pickers/TagPicker.CustomRemoveIcon.Example.tsx') as string;
+const TagPickerExampleCode =
+  require('!raw-loader?esModule=false!@fluentui/react-examples/src/react/Pickers/TagPicker.Basic.Example.tsx') as string;
+const TagPickerInlineExampleCode =
+  require('!raw-loader?esModule=false!@fluentui/react-examples/src/react/Pickers/TagPicker.Inline.Example.tsx') as string;
+const PickerCustomResultExampleCode =
+  require('!raw-loader?esModule=false!@fluentui/react-examples/src/react/Pickers/Picker.CustomResult.Example.tsx') as string;
+const TagPickerCustomRemoveIconExampleCode =
+  require('!raw-loader?esModule=false!@fluentui/react-examples/src/react/Pickers/TagPicker.CustomRemoveIcon.Example.tsx') as string;
 
 export const PickersPageProps: IDocPageProps = {
   title: 'Pickers',