+++ packages/react-window-provider/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log - @fluentui/react-window-provider
 
-This log was last generated on Fri, 03 Feb 2023 07:50:06 GMT and should not be manually modified.
+This log was last generated on Wed, 01 Mar 2023 07:45:41 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [2.2.7](https://github.com/microsoft/fluentui/tree/@fluentui/react-window-provider_v2.2.7)
+
+Wed, 01 Mar 2023 07:45:41 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-window-provider_v2.2.6..@fluentui/react-window-provider_v2.2.7)
+
+### Patches
+
+- Bump @fluentui/test-utilities to v8.3.0 ([PR #26980](https://github.com/microsoft/fluentui/pull/26980) by beachball)
 
 ## [2.2.6](https://github.com/microsoft/fluentui/tree/@fluentui/react-window-provider_v2.2.6)