+++ packages/scheme-utilities/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log - @fluentui/scheme-utilities
 
-This log was last generated on Sat, 18 Feb 2023 01:37:02 GMT and should not be manually modified.
+This log was last generated on Wed, 01 Mar 2023 07:45:41 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [8.3.25](https://github.com/microsoft/fluentui/tree/@fluentui/scheme-utilities_v8.3.25)
+
+Wed, 01 Mar 2023 07:45:41 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/scheme-utilities_v8.3.24..@fluentui/scheme-utilities_v8.3.25)
+
+### Patches
+
+- Bump @fluentui/theme to v2.6.24 ([PR #26980](https://github.com/microsoft/fluentui/pull/26980) by beachball)
 
 ## [8.3.24](https://github.com/microsoft/fluentui/tree/@fluentui/scheme-utilities_v8.3.24)